\name{triangle}
\docType{data}
\alias{triangle}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Simulated functional bivariate dataset.
}
\description{
A simulated dataset of bivariate functional data from 3 groups.
}
\usage{
triangle
}
%- maybe also 'usage' for other objects documented here.
\format{
A dataframe with 100 lines and 203 columns. A row represents a curve and a column a sampling point.
Columns from 1 to 101 are sampling points of the first functional variable.
Columns from 102 to 202 are sampling points of the second functional variable.
Column 203 is the variable which indicates the true clustering partition.
}

