\name{trigo}
\docType{data}
\alias{trigo}
\title{
  Simulated functional univariate dataset
}
\description{
  A simulated dataset of univariate functional data from 2 groups.
}
\usage{
  trigo
}
\format{
  A dataframe with 100 lines and 101 columns. A row represents a curve and a column a sampling point.
  Columns from 1 to 100 are sampling points of the functional variable.
  Column 101 is the variable which indicates the true clustering partition.}


